+++ packages/symfony-bundle/tests/DependencyInjection/ConfigurationTest.php
@@ -113,4 +113,133 @@ class ConfigurationTest extends TestCase
         $this->assertFalse($config['singleFileMode']);
         $this->assertSame('bundle.ts', $config['singleFileOutput']);
     }
+
+    public function testInterfacesMapSyntax(): void
+    {
+        $configuration = new Configuration();
+        $processor = new Processor();
+        $config = $processor->processConfiguration($configuration, [
+            [
+                'interfaces' => [
+                    'vendor/foo/src/Dto.php' => [
+                        'output' => 'external/foo/',
+                    ],
+                    'vendor/bar/src/Model.php' => [
+                        'output' => 'external/bar/',
+                    ],
+                ],
+            ]
+        ]);
+
+        $this->assertCount(2, $config['interfaces']);
+        $this->assertSame('vendor/foo/src/Dto.php', $config['interfaces'][0]['path']);
+        $this->assertSame('external/foo/', $config['interfaces'][0]['output']);
+        $this->assertSame('vendor/bar/src/Model.php', $config['interfaces'][1]['path']);
+        $this->assertSame('external/bar/', $config['interfaces'][1]['output']);
+    }
+
+    public function testInterfacesListSyntax(): void
+    {
+        $configuration = new Configuration();
+        $processor = new Processor();
+        $config = $processor->processConfiguration($configuration, [
+            [
+                'interfaces' => [
+                    [
+                        'path' => 'vendor/foo/src/Dto.php',
+                        'output' => 'external/foo/',
+                    ],
+                    [
+                        'path' => 'vendor/bar/src/Model.php',
+                        'output' => 'external/bar/',
+                    ],
+                ],
+            ]
+        ]);
+
+        $this->assertCount(2, $config['interfaces']);
+        $this->assertSame('vendor/foo/src/Dto.php', $config['interfaces'][0]['path']);
+        $this->assertSame('external/foo/', $config['interfaces'][0]['output']);
+        $this->assertSame('vendor/bar/src/Model.php', $config['interfaces'][1]['path']);
+        $this->assertSame('external/bar/', $config['interfaces'][1]['output']);
+    }
+
+    public function testDirectoriesMapSyntax(): void
+    {
+        $configuration = new Configuration();
+        $processor = new Processor();
+        $config = $processor->processConfiguration($configuration, [
+            [
+                'directories' => [
+                    'vendor/foo/src/' => [
+                        'output' => 'external/foo/',
+                        'requireAnnotation' => true,
+                    ],
+                    'vendor/bar/src/' => [
+                        'output' => 'external/bar/',
+                        'requireAnnotation' => false,
+                    ],
+                ],
+            ]
+        ]);
+
+        $this->assertCount(2, $config['directories']);
+        $this->assertSame('vendor/foo/src/', $config['directories'][0]['path']);
+        $this->assertSame('external/foo/', $config['directories'][0]['output']);
+        $this->assertTrue($config['directories'][0]['requireAnnotation']);
+        $this->assertSame('vendor/bar/src/', $config['directories'][1]['path']);
+        $this->assertSame('external/bar/', $config['directories'][1]['output']);
+        $this->assertFalse($config['directories'][1]['requireAnnotation']);
+    }
+
+    public function testDirectoriesListSyntax(): void
+    {
+        $configuration = new Configuration();
+        $processor = new Processor();
+        $config = $processor->processConfiguration($configuration, [
+            [
+                'directories' => [
+                    [
+                        'path' => 'vendor/foo/src/',
+                        'output' => 'external/foo/',
+                        'requireAnnotation' => true,
+                    ],
+                    [
+                        'path' => 'vendor/bar/src/',
+                        'output' => 'external/bar/',
+                        'requireAnnotation' => false,
+                    ],
+                ],
+            ]
+        ]);
+
+        $this->assertCount(2, $config['directories']);
+        $this->assertSame('vendor/foo/src/', $config['directories'][0]['path']);
+        $this->assertSame('external/foo/', $config['directories'][0]['output']);
+        $this->assertTrue($config['directories'][0]['requireAnnotation']);
+        $this->assertSame('vendor/bar/src/', $config['directories'][1]['path']);
+        $this->assertSame('external/bar/', $config['directories'][1]['output']);
+        $this->assertFalse($config['directories'][1]['requireAnnotation']);
+    }
+
+    public function testDirectoriesListSyntaxWithDefaultRequireAnnotation(): void
+    {
+        $configuration = new Configuration();
+        $processor = new Processor();
+        $config = $processor->processConfiguration($configuration, [
+            [
+                'directories' => [
+                    [
+                        'path' => 'vendor/foo/src/',
+                        'output' => 'external/foo/',
+                    ],
+                ],
+            ]
+        ]);
+
+        $this->assertCount(1, $config['directories']);
+        $this->assertSame('vendor/foo/src/', $config['directories'][0]['path']);
+        $this->assertSame('external/foo/', $config['directories'][0]['output']);
+        $this->assertFalse($config['directories'][0]['requireAnnotation']);
+    }
 }