+++ packages/symfony-bundle/tests/Fixtures/ExternalNoAttribute.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paneon\PhpToTypeScriptBundle\Tests\Fixtures;
+
+class ExternalNoAttribute
+{
+    public string $foo;
+}